Done: -Move recommendations logic to Recommnedation service -Add relations between like, ban to profile -Move some like & ban logic from controller to model -Rename action in recommendation logic

+++ database/migrations/2019_08_28_125533_create_likes_table.php
@@ -15,7 +15,7 @@ class CreateLikesTable extends Migration
     {
         Schema::create('likes', function (Blueprint $table) {
             $table->bigIncrements('like_id');
-            $table->bigInteger('profile_id');
+            $table->unsignedInteger('profile_id');
             $table->bigInteger('partner_id');
             $table->timestamps();
         });
